// packages/theme-engine/src/tokens.d.ts
export declare const tokens: {
    colors: {
        primary: string;
        secondary: string;
        background: string;
        text: string;
    };
    spacing: {
        xs: string;
        sm: string;
        md: string;
        lg: string;
        xl: string;
    };
    typography: {
        fonts: {
            sans: string;
            mono: string;
        };
        sizes: {
            xs: string;
            sm: string;
            md: string;
            lg: string;
            xl: string;
            '2xl': string;
            '3xl': string;
            '4xl': string;
            '5xl': string;
        };
    };
};
//# sourceMappingURL=tokens.d.ts.map